requirements.txt
opencv-python>=4.8.0
numpy>=1.26.0
PyQt6>=6.5.0
scipy>=1.11.0
scikit-image>=0.21.0
matplotlib>=3.7.0
Pillow>=10.0.0
